AiTranslation-server/requirements.txt
Django>=5.2.3
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.0
requests>=2.31.0
websocket-client>=1.6.4
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.12.0
python-dotenv>=1.0.0
redis>=5.0.0
celery>=5.3.0
oss2>=2.18.0 